Rename utils to "db"
