+++ cmd/sonictool/db/dbutils.go
@@ -1,4 +1,4 @@
-package utils
+package db
 
 import (
 	"errors"
